Refactoring was done: added header to Makefile, source code more structure for impore readability
